#163 extract domain validators and mapping helpers; refactor level.ts as facade
